// web/shared/enums/routes.enum.ts
export enum ROUTES {
    LOGIN = '/login',
    DASHBOARD = '/dashboard',
    CONSULTAR = '/consultar',
    INSERIR = '/inserir',
    EXCLUIR = '/excluir',
    USUARIOS = '/usuarios',
    CONTESTACAO = '/contestacao',
    ADMINISTRADOR = '/administrador',
    TERMOS_USO = '/termos-de-uso',
    POLITICA_PRIVACIDADE = '/politica-privacidade'
}